feat: Add Accessible Interactive Components + Example Usage
Added some accessible UI elements with some throw-away styling. These include a base Button element, form controls / inputs, a Modal dialog container, and a Link wrapper which supports external links.

An example of how to use some of these elements has been added to the Home route.

The Form and Group components are designed to automatically wire up props to child elements to reduce the boilerplate typically needed to use the base components that come from Reakit.

Also fixes an error with the commit message hook whereby it was improperly loading the config.